Moved string_with_no_placeholders_in_single_quotes
